+++ R/termSim.R
@@ -6,7 +6,7 @@
 ##'
 ##'@param t1 term vector
 ##'@param t2 term vector
-##'@param method one of "Wang", "Resnik", "Rel", "Jiang", and "Lin".
+##'@param method one of "Wang", "Resnik", "Rel", "Jiang", and "Lin", "TCSS".
 ##'@param semData GOSemSimDATA object
 ##'@return score matrix
 ##'@export
@@ -14,13 +14,13 @@
 termSim <- function(t1,
                     t2,
                     semData,
-                    method=c("Wang","Resnik","Rel","Jiang","Lin")
+                    method=c("Wang","Resnik","Rel","Jiang","Lin", "TCSS")
                     ) {
 
     method <- match.arg(method)
 
     if (all(is.na(t1)) || all(is.na(t2)))
-        return (NA)
+        return(NA)
 
     t1 <- t1[!is.na(t1)]
     t2 <- t2[!is.na(t2)]
@@ -32,8 +32,10 @@ termSim <- function(t1,
     ## t2 <- unique(t2)
 
     if ( method %in% c("Resnik", "Jiang", "Lin", "Rel") ) {
-        return(infoContentMethod(t1, t2, method=method, semData))
+        return(infoContentMethod(t1, t2, method = method, semData))
     } else if ( method == "Wang" ) {
         return(wangMethod(t1, t2, semData@ont))
+    } else if ( method == "TCSS" ) {
+        return(tcssMethod(t1, t2, semData))
     }
 }