+++ NEWS.md
@@ -1,7 +1,54 @@
+# GOSemSim 2.21.1
+
++ Avoid eval-parse in `load_OrgDb()` (2022-01-10, Mon)
+
+# GOSemSim 2.20
+
++ Bioconductor 3.14 release
+
+# GOSemSim 2.19.1
+
++ TCSS method (@qibaiqi, #35; 2021-08-02, Mon)
+
++ GOSemSim 2.18.0
+
++ Bioconductor 3.13 release
+
+# GOSemSim 2.17.1
+
++ bug fixed according to the update of GO.db (2020-10-29, Thu)
+  - <https://github.com/YuLab-SMU/GOSemSim/issues/32>
+  
+# GOSemSim 2.16.0
+
++ Bioconductor 3.12 release (2020-10-28, Wed)
+
+# GOSemSim 2.15.2
+
++ new site, <https://yulab-smu.top/biomedical-knowledge-mining-book/> for documentation (2020-09-04, Fri)
++ update vignette
++ update `data/gotbl`
+
+# GOSemSim 2.15.1
+
++ bug fixed of IC method when input IDs contain invalid terms. (2020-07-25, Sat)
+
+# GOSemSim 2.14.0
+
++ Bioconductor 3.11 release
+
+
+# GOSemSim 2.13.1
+
++ add new citation (2020-03-19, Thu)
++ fixed compiling error due to the change of Rcpp 
+  - <https://github.com/YuLab-SMU/GOSemSim/issues/27>
+
 # GOSemSim 2.7.1
 
 + `mgeneSim` and `mclusterSim` now always return matrix (2018-08-08, Wed)
     - <https://www.biostars.org/p/330642/#331598>
+
 # GOSemSim 2.5.1
 
 + return NA for deprecated IDs (2018-01-09, Fri)